abstract MCWorld content + working on ChunkManager for server module
