+++ pizzaserver-server/src/main/java/io/github/willqi/pizzaserver/server/world/World.java
@@ -2,6 +2,7 @@ package io.github.willqi.pizzaserver.server.world;
 
 import io.github.willqi.pizzaserver.mcworld.MCWorld;
 import io.github.willqi.pizzaserver.server.Server;
+import io.github.willqi.pizzaserver.server.world.chunks.ChunkManager;
 
 import java.io.Closeable;
 import java.io.File;
@@ -10,7 +11,10 @@ import java.io.IOException;
 public class World implements Closeable {
 
     private final Server server;
+
     private final MCWorld mcWorld;
+
+    private final ChunkManager chunkManager = new ChunkManager(this);
 
 
     public World(Server server, File worldDirectory) throws IOException {
@@ -22,9 +26,13 @@ public class World implements Closeable {
         return this.server;
     }
 
+    public ChunkManager getChunkManager() {
+        return this.chunkManager;
+    }
+
     @Override
     public void close() throws IOException {
-        this.mcWorld.close();
+
     }
 
 }
