+++ pizzaserver-server/src/main/java/io/github/willqi/pizzaserver/server/world/chunks/Chunk.java
@@ -1,0 +1,9 @@
+package io.github.willqi.pizzaserver.server.world.chunks;
+
+import io.github.willqi.pizzaserver.server.player.Player;
+
+public class Chunk {
+
+
+
+}
